+++ src/entity/area.rs
@@ -13,7 +13,11 @@ use serde::{Deserialize, Serialize};
 /// Areas are historical and existing geographic regions. Areas include countries, sub-divisions,
 /// counties, municipalities, cities, districts and islands.
 #[derive(Debug, Serialize, Deserialize, PartialEq, Clone, Default)]
-#[serde(rename_all(deserialize = "kebab-case"))]
+#[cfg_attr(
+    feature = "legacy_serialize",
+    serde(rename_all(deserialize = "kebab-case"))
+)]
+#[cfg_attr(not(feature = "legacy_serialize"), serde(rename_all = "kebab-case"))]
 #[serde(default)]
 pub struct Area {
     /// See [MusicBrainz Identifier](https://musicbrainz.org/doc/MusicBrainz_Identifier).