+++ src/entity/artist.rs
@@ -20,7 +20,11 @@ use serde::{Deserialize, Serialize};
 /// (like a photographer, an illustrator, or a poet whose writings are set to music), or even a
 /// fictional character. For some other special cases, see special purpose artists.
 #[derive(Debug, Serialize, Deserialize, PartialEq, Clone, Default)]
-#[serde(rename_all(deserialize = "kebab-case"))]
+#[cfg_attr(
+    feature = "legacy_serialize",
+    serde(rename_all(deserialize = "kebab-case"))
+)]
+#[cfg_attr(not(feature = "legacy_serialize"), serde(rename_all = "kebab-case"))]
 #[serde(default)]
 pub struct Artist {
     /// See [MusicBrainz Identifier](https://musicbrainz.org/doc/MusicBrainz_Identifier).
@@ -142,10 +146,11 @@ pub enum ArtistType {
 /// changes are only reflected in the DB, not in actual MB code.
 /// Variants are derived from the `gender` table in the MusicBrainz database.
 #[non_exhaustive]
-#[derive(Debug, Serialize, Deserialize, PartialEq, Eq, Clone)]
+#[derive(Debug, Serialize, Deserialize, PartialEq, Eq, Clone, Default)]
 pub enum Gender {
     Male,
     Female,
+    #[default]
     Other,
     /// For cases where gender just doesn't apply at all (like companies entered as artists).
     #[serde(rename = "Not applicable")]
@@ -154,12 +159,6 @@ pub enum Gender {
     /// If you ever see a `Gender::UnrecognizedGender` in the wild, let us know and file an issue/pull request!
     #[serde(other)]
     UnrecognizedGender,
-}
-
-impl Default for Gender {
-    fn default() -> Self {
-        Gender::Other
-    }
 }
 
 #[derive(Debug, QueryBuilder, Default)]
@@ -222,6 +221,7 @@ impl_includes!(
     Artist,
     (with_recordings, Include::Subquery(Subquery::Recordings)),
     (with_releases, Include::Subquery(Subquery::Releases)),
+    (with_medias, Include::Subquery(Subquery::Media)),
     (
         with_releases_and_discids,
         Include::Subquery(Subquery::ReleasesWithDiscIds)