+++ src/entity/event.rs
@@ -51,7 +51,11 @@ pub enum EventType {
 /// An event refers to an organised event which people can attend, and is relevant to MusicBrainz.
 /// Generally this means live performances, like concerts and festivals.
 #[derive(Debug, Serialize, Deserialize, PartialEq, Clone)]
-#[serde(rename_all(deserialize = "kebab-case"))]
+#[cfg_attr(
+    feature = "legacy_serialize",
+    serde(rename_all(deserialize = "kebab-case"))
+)]
+#[cfg_attr(not(feature = "legacy_serialize"), serde(rename_all = "kebab-case"))]
 pub struct Event {
     /// See [MusicBrainz Identifier](https://musicbrainz.org/doc/MusicBrainz_Identifier).
     pub id: String,
@@ -68,7 +72,7 @@ pub struct Event {
     pub cancelled: Option<bool>,
 
     /// The time is the start time of the event.
-    pub time: String,
+    pub time: Option<String>,
 
     /// The setlist stores a list of songs performed, optionally including links to artists and works.
     /// See the setlist documentation for syntax and examples.