+++ src/entity/genre.rs
@@ -4,8 +4,14 @@ use serde::{Deserialize, Serialize};
 /// See [Genre](https://musicbrainz.org/doc/Genre) and
 /// [supported genres](https://musicbrainz.org/genres) for more information.
 #[derive(Debug, Serialize, Deserialize, PartialEq, Eq, Clone)]
-#[serde(rename_all(deserialize = "kebab-case"))]
+#[cfg_attr(
+    feature = "legacy_serialize",
+    serde(rename_all(deserialize = "kebab-case"))
+)]
+#[cfg_attr(not(feature = "legacy_serialize"), serde(rename_all = "kebab-case"))]
 pub struct Genre {
-    pub count: u32,
+    pub id: Option<String>,
+    pub count: Option<u32>,
     pub name: String,
+    pub disambiguation: Option<String>,
 }