+++ src/entity/instrument.rs
@@ -14,7 +14,11 @@ use lucene_query_builder::QueryBuilder;
 /// can only be edited by relationship editors.
 /// See [Instrument List](https://musicbrainz.org/instruments) for the full list.
 #[derive(Debug, Serialize, Deserialize, PartialEq, Clone)]
-#[serde(rename_all(deserialize = "kebab-case"))]
+#[cfg_attr(
+    feature = "legacy_serialize",
+    serde(rename_all(deserialize = "kebab-case"))
+)]
+#[cfg_attr(not(feature = "legacy_serialize"), serde(rename_all = "kebab-case"))]
 pub struct Instrument {
     /// See [MusicBrainz Identifier](https://musicbrainz.org/doc/MusicBrainz_Identifier).
     pub id: String,