+++ src/entity/tag.rs
@@ -3,5 +3,6 @@ use serde::{Deserialize, Serialize};
 #[derive(Debug, Serialize, Deserialize, PartialEq, Eq, Clone)]
 pub struct Tag {
     pub name: String,
-    pub count: i32,
+    pub count: Option<i32>,
+    pub score: Option<i32>,
 }