+++ tests/blocking_tests/release/release_includes.rs
@@ -131,7 +131,15 @@ fn should_get_release_level_relations() {
     let mut target_types = HashSet::new();
     for media in become_desert.media.unwrap().iter() {
         for track in media.tracks.as_ref().unwrap().iter() {
-            for relation in track.recording.relations.as_ref().unwrap().iter() {
+            for relation in track
+                .recording
+                .as_ref()
+                .unwrap()
+                .relations
+                .as_ref()
+                .unwrap()
+                .iter()
+            {
                 target_types.insert(relation.target_type.clone());
             }
         }